2003-07-28  Piers Haken	<piersh@friskit.com>
	* DefaultContext.cs: count() should return a double
	* Expression.cs: better support for handling non-double numbers

svn path=/trunk/mcs/; revision=16799

+++ mcs/class/System.XML/System.Xml.XPath/DefaultContext.cs
@@ -240,7 +240,7 @@ namespace System.Xml.XPath
 		public override XPathResultType [] ArgTypes { get { return new XPathResultType [] { XPathResultType.NodeSet }; }}
 		public override object TypesafeInvoke (XsltContext xsltContext, object[] args, XPathNavigator docContext)
 		{
-			return ((BaseIterator) args [0]).Count;
+			return (double) ((BaseIterator) args [0]).Count;
 		}
 		public override string Name { get { return "count"; }}
 	}
